Final Finishing Touches before Final Presentation.
[NEW] AutoQPGen :: /Assets/QuestionBank_FORMAT/*.pdf :: Added two new question banks for demo purposes that have been pre-tested to work correctly regarding data extraction.
[NEW] AutoQPGen :: /Assets/QuestionBank_FORMAT/*.docx :: Added source files (DOCX) for the question banks that are newly added for demo purposes.
[NEW] AutoQPGen :: /Assets/QuestionBank_FORMAT/Template/QBank-TEMPLATE.pdf :: Added a new demo template clearly showcasing the accepted new format for question banks that brings increased readbility for both human and AutoQPGen.
[NEW] AutoQPGen :: /Assets/QuestionBank_FORMAT/Template/QBank-TEMPLATE.docx :: Added source file (DOCX) for the demo template question bank.
[NEW] AutoQPGen :: /models/QScanEngine.py :: Introducing game-changing PDF Extraction library called PDFPlumber which employs state-of-the-art algorithms to extract data that is readable from pdf without redundant white spaces or any such unwanted characters and mainly focusing on "STRUCTURED READING" of PDF content just like human readable way.

[FIX] AutoQPGen :: /app.py :: Fixed a fatal error that caused the web app to crash while purging redundant documents.
[FIX] AutoQPGen :: /Assets/TestQBanks/ :: This folder has been removed as result of moving to new format and the older question bank format is unsupported.
[FIX] AutoQPGen :: /models/QScanEngine.py :: Replaced the FITZ data extraction library with the new PDFPlumber library.
[FIX] AutoQPGen :: /models/QScanEngine.py :: Redesigned questionSetter() function to properly sort the operated questions before display and group questions based on Ascending order of Modules.
[FIX] AutoQPGen :: /models/QScanEngine.py :: Rerouted the function extract_text_from_pdf() to now get data using PDFPlumber reflecting new format acceptance.
[FIX] AutoQPGen :: /models/QScanEngine.py :: Reconfigured SpaCy model to better figure out Subject data from question bank.
[FIX] AutoQPGen :: /README.md :: Removed fitz/PyMuPDF dependency issue section as the library is itself no longer used.
[FIX] AutoQPGen :: /README.md :: Added a (PDF Structured Extraction) to the Technologies used section.
[FIX] AutoQPGen :: /README.md :: Fixed typos in setup instructions.
[FIX] AutoQPGen :: /requirements.txt :: Added libraries {PDFPlumber, GoogleGenAI} and removed libraries {PDFkit, fitz, PyPDF2, secrets}.

+++ app.py
@@ -90,9 +90,9 @@ def deletePaperRemains(paperId):
     generatedDocxFolder = './static/GeneratedDocx'
     docxFileFound = search_for_file(generatedDocxFolder, paperId, '.docx')
     pdfFileFound = search_for_file(generatedPapersFolder, paperId, '.pdf')
-    if docxFileFound is not None:
+    if docxFileFound is not None and os.path.exists(docxFileFound):
         os.remove(docxFileFound)
-    if pdfFileFound is not None:
+    if pdfFileFound is not None and os.path.exists(pdfFileFound):
         os.remove(pdfFileFound)
     eventLogger(f'DOCX and PDF Remains of PaperID {paperId} were cleared.')
